model.py
import numpy as np
from PIL import Image
import io
import cv2

class WasteClassifier:
    """
    A classifier that uses image features for waste classification
    """
    def __init__(self):
        self.categories = ['plastic', 'glass', 'metal', 'paper', 'organic', 'e-waste']
    
    def predict(self, image_array):
        """
        Predict based on image features like color, texture, and shape
        """
        # Remove batch dimension if present
        if len(image_array.shape) == 4:
            image_array = image_array[0]
        
        # Convert to OpenCV format for better feature extraction
        img_cv = (image_array * 255).astype(np.uint8)
        
        # Calculate average color values per channel
        avg_colors = np.mean(image_array, axis=(0, 1))
        r, g, b = avg_colors
        
        # Calculate color variance (texture)
        color_variance = np.var(image_array, axis=(0, 1))
        texture_score = np.mean(color_variance)
        
        # Calculate brightness
        brightness = np.mean(image_array)
        
        # Calculate color ratios
        red_ratio = r / (r + g + b + 1e-6)
        green_ratio = g / (r + g + b + 1e-6)
        blue_ratio = b / (r + g + b + 1e-6)
        
        # Calculate edge detection for shape analysis
        gray = cv2.cvtColor(img_cv, cv2.COLOR_RGB2GRAY)
        edges = cv2.Canny(gray, 100, 200)
        edge_density = np.mean(edges > 0)
        
        # Calculate color saturation
        hsv = cv2.cvtColor(img_cv, cv2.COLOR_RGB2HSV)
        saturation = np.mean(hsv[:, :, 1]) / 255.0
        
        # Initialize probabilities
        probs = np.zeros(len(self.categories))
        
        # Plastic detection (shiny, smooth surface, often colorful)
        if (saturation > 0.3 and  # Colorful
            edge_density < 0.1 and  # Smooth surface
            texture_score < 0.2 and  # Low texture
            brightness > 0.5):  # Shiny
            probs[0] += 0.9
        
        # Glass detection (transparent, reflective, sharp edges)
        elif (edge_density > 0.2 and  # Sharp edges
              saturation < 0.2 and  # Low color saturation
              brightness > 0.7 and  # Very bright
              texture_score < 0.1):  # Very smooth
            probs[1] += 0.9
        
        # Metal detection (reflective, high contrast, sharp edges)
        elif (edge_density > 0.15 and  # Sharp edges
              brightness > 0.6 and  # Bright
              texture_score > 0.3 and  # High texture
              saturation < 0.3):  # Low color saturation
            probs[2] += 0.9
        
        # Paper detection (flat surface, high brightness, low texture)
        elif (brightness > 0.8 and  # Very bright
              texture_score < 0.1 and  # Very smooth
              edge_density < 0.05 and  # Few edges
              saturation < 0.2):  # Low color saturation
            probs[3] += 0.9
        
        # Organic detection (textured, natural colors, irregular shape)
        elif (texture_score > 0.4 and  # High texture
              green_ratio > 0.3 and  # Natural colors
              edge_density > 0.1 and  # Irregular shape
              saturation > 0.2):  # Moderate color saturation
            probs[4] += 0.9
        
        # E-waste detection (complex shape, mixed materials, high contrast)
        elif (edge_density > 0.25 and  # Complex shape
              texture_score > 0.3 and  # Mixed materials
              brightness < 0.5 and  # Dark
              saturation > 0.2):  # Some color
            probs[5] += 0.9
        
        # Add some randomness to make it more realistic
        probs += np.random.random(len(probs)) * 0.1
        
        # Normalize probabilities
        probs = probs / np.sum(probs)
        
        return probs

def load_model():
    """
    Load the waste classification model
    """
    model = WasteClassifier()
    return model

def preprocess_image(image):
    """
    Preprocess the image to match the model's expected input
    
    Args:
        image: PIL Image or file path
    
    Returns:
        Preprocessed image as numpy array
    """
    if isinstance(image, str):
        image = Image.open(image)
    
    # Convert to RGB if needed
    if image.mode != 'RGB':
        image = image.convert('RGB')
    
    # Resize to expected dimensions
    image = image.resize((224, 224))
    
    # Convert to numpy array and normalize
    img_array = np.array(image) / 255.0
    
    # Add batch dimension
    img_array = np.expand_dims(img_array, axis=0)
    
    return img_array

def predict_waste_class(model, image_array):
    """
    Predict waste class from image array
    
    Args:
        model: WasteClassifier instance
        image_array: Preprocessed image as numpy array
    
    Returns:
        Tuple of (predicted_class, confidence)
    """
    # Get model prediction
    predictions = model.predict(image_array)
    
    # Get the index of the highest probability
    predicted_index = np.argmax(predictions)
    
    # Get the class name and confidence
    predicted_class = model.categories[predicted_index]
    confidence = predictions[predicted_index] * 100
    
    return predicted_class, confidence
