requirements.txt
streamlit==1.31.1
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
plotly==5.19.0
matplotlib==3.8.3
flask==3.0.2 